Add CLI data dictionary formatter
Currently, if a data dictionary file that is exported from REDCap is
edited in a third-party program, the carriage returns in the fields get
changed to newlines. As far as I can tell, REDCap only contains carriage
returns within fields, so this script changes such newlines back.

(This seems like a simple change, but it will cure many a headache.)