Change failed-date record IDs procedure to always correcting them
The alternative is to transmit with --all, then correct in the Imported
from Laptops project for each form and set an exception in
special_cases.yml for this record.

This slightly violates the principle that all correction happens in the
Import project, but the benefit is not having another list of exceptions
to maintain.